% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer_w_units.R
\name{get_answer_w_units}
\alias{get_answer_w_units}
\title{Extract answers and corresponding units}
\usage{
get_answer_w_units(results)
}
\arguments{
\item{results}{element from the 'results' list from the output of xml_get_answers}
}
\description{
Gets answers and units from results part of xml grading reports
}
