% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_xml.R
\name{get_data_from_xml}
\alias{get_data_from_xml}
\title{Extract data from xml document}
\usage{
get_data_from_xml(XML_document, primary = FALSE, n_head = NA)
}
\arguments{
\item{XML_document}{the xml_document part of a read_xml output from which data is to be extracted}

\item{primary}{logical; are we reading primary grading results?}

\item{n_head}{limit output to first n_head grading reports}
}
\description{
From an xml document, get relevant characteristics and a list of results. Designed to work on the output of 'xml_to_tibble'.
}
