% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_get_answers.R
\name{xml_get_answers}
\alias{xml_get_answers}
\title{Unfold answers from data}
\usage{
xml_get_answers(tib0, show_progress = FALSE)
}
\arguments{
\item{tib0}{result of get_data_from_xml}

\item{show_progress}{logical; if FALSE (default), progress bar is not shown}
}
\description{
The output from 'get_data_from_xml' includes a list column with xml_nodesets called 'Results'. This
column has all the information about questions and answers.
'xml_get_answers' extracts the 'Question' and 'Answer' information from said column,
using the function 'get_answer_w_units'.
}
